refactor(pod): move pod functions to a new file
this allows us to grow context within the same package
